requirements.txt
Flask==3.0.3
numpy==1.24.1
Pillow==11.0.0
torch==2.2.2
torchvision==0.17.2
gunicorn
google-generativeai==0.8.3
albumentations==1.2.1
opencv-python



